+++ src/adc/reader/classification/_subdir.py
@@ -5,10 +5,10 @@ from typing import List, Iterable, Union
 from wai.logging import LOGGING_WARNING
 from kasperl.api import Reader
 from adc.api import AudioClassificationData
-from seppl.placeholders import PlaceholderSupporter, placeholder_list
+from seppl.variables import VariableSupporter, variable_list
 
 
-class SubDirAudioClassificationReader(Reader, PlaceholderSupporter):
+class SubDirAudioClassificationReader(Reader, VariableSupporter):
 
     def __init__(self, source: Union[str, List[str]] = None, source_list: Union[str, List[str]] = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
@@ -53,8 +53,8 @@ class SubDirAudioClassificationReader(Reader, PlaceholderSupporter):
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-i", "--input", type=str, help="Path to the directory with the sub-directories containing the audio files; " + placeholder_list(obj=self), required=False, nargs="*")
-        parser.add_argument("-I", "--input_list", type=str, help="Path to the text file(s) listing the directories to use; " + placeholder_list(obj=self), required=False, nargs="*")
+        parser.add_argument("-i", "--input", type=str, help="Path to the directory with the sub-directories containing the audio files; " + variable_list(obj=self), required=False, nargs="*")
+        parser.add_argument("-I", "--input_list", type=str, help="Path to the text file(s) listing the directories to use; " + variable_list(obj=self), required=False, nargs="*")
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -114,7 +114,7 @@ class SubDirAudioClassificationReader(Reader, PlaceholderSupporter):
         # find all subdirs
         self._sub_dirs = dict()
         for input_dir in all_dirs:
-            input_dir = self.session.expand_placeholders(input_dir)
+            input_dir = self.session.expand_variables(input_dir)
             if not os.path.exists(input_dir):
                 self.logger().warning("Directory does not exist: %s" % input_dir)
                 continue