+++ src/adc/reader/speech/_festvox.py
@@ -4,7 +4,7 @@ import re
 from typing import List, Iterable, Union
 
 from seppl.io import locate_files
-from seppl.placeholders import PlaceholderSupporter, placeholder_list
+from seppl.variables import VariableSupporter, variable_list
 from wai.logging import LOGGING_WARNING
 
 from kasperl.api import Reader
@@ -17,7 +17,7 @@ LINE_REGEX = '^\\( (?P<filename>.*) "(?P<transcription>.*)" \\)$'
 LINE_PATTERN = re.compile(LINE_REGEX)
 
 
-class FestVoxSpeechReader(Reader, PlaceholderSupporter):
+class FestVoxSpeechReader(Reader, VariableSupporter):
 
     def __init__(self, source: Union[str, List[str]] = None, source_list: Union[str, List[str]] = None,
                  rel_path: str = None, resume_from: str = None,
@@ -70,8 +70,8 @@ class FestVoxSpeechReader(Reader, PlaceholderSupporter):
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-i", "--input", type=str, help="Path to the text file(s) to read; glob syntax is supported; " + placeholder_list(obj=self), required=False, nargs="*")
-        parser.add_argument("-I", "--input_list", type=str, help="Path to the text file(s) listing the text files to use; " + placeholder_list(obj=self), required=False, nargs="*")
+        parser.add_argument("-i", "--input", type=str, help="Path to the text file(s) to read; glob syntax is supported; " + variable_list(obj=self), required=False, nargs="*")
+        parser.add_argument("-I", "--input_list", type=str, help="Path to the text file(s) listing the text files to use; " + variable_list(obj=self), required=False, nargs="*")
         parser.add_argument("--resume_from", type=str, help="Glob expression matching the file to resume from, e.g., '*/012345.txt'", required=False)
         parser.add_argument("-r", "--rel_path", type=str, help="The relative path to the audio files.", required=False, default=".")
         return parser