+++ src/adc/reader/speech/_hf_audiofolder.py
@@ -5,7 +5,7 @@ from typing import List, Iterable, Union
 
 from wai.logging import LOGGING_WARNING
 from seppl.io import locate_files
-from seppl.placeholders import PlaceholderSupporter, placeholder_list
+from seppl.variables import VariableSupporter, variable_list
 
 from kasperl.api import Reader
 from adc.api import SpeechData
@@ -13,7 +13,7 @@ from adc.api import SpeechData
 HF_AUDIOFOLDER_EXPECTED_HEADER = "file_name,transcription"
 
 
-class HuggingFaceAudioFolderSpeechReader(Reader, PlaceholderSupporter):
+class HuggingFaceAudioFolderSpeechReader(Reader, VariableSupporter):
 
     def __init__(self, source: Union[str, List[str]] = None, source_list: Union[str, List[str]] = None,
                  resume_from: str = None, logger_name: str = None, logging_level: str = LOGGING_WARNING):
@@ -62,8 +62,8 @@ class HuggingFaceAudioFolderSpeechReader(Reader, PlaceholderSupporter):
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-i", "--input", type=str, help="Path to the CSV file(s) to read; glob syntax is supported; " + placeholder_list(obj=self), required=False, nargs="*")
-        parser.add_argument("-I", "--input_list", type=str, help="Path to the text file(s) listing the CSV files to use; " + placeholder_list(obj=self), required=False, nargs="*")
+        parser.add_argument("-i", "--input", type=str, help="Path to the CSV file(s) to read; glob syntax is supported; " + variable_list(obj=self), required=False, nargs="*")
+        parser.add_argument("-I", "--input_list", type=str, help="Path to the text file(s) listing the CSV files to use; " + variable_list(obj=self), required=False, nargs="*")
         parser.add_argument("--resume_from", type=str, help="Glob expression matching the file to resume from, e.g., '*/012345.csv'", required=False)
         return parser