+++ src/adc/writer/classification/_adams.py
@@ -4,12 +4,12 @@ from typing import List
 
 from wai.logging import LOGGING_WARNING
 from wai.common.file.report import Report, Field, save
-from seppl.placeholders import placeholder_list, InputBasedPlaceholderSupporter
+from seppl.variables import InputBasedVariableSupporter, variable_list
 from kasperl.api import SplittableStreamWriter, make_list, AnnotationsOnlyWriter, add_annotations_only_writer_param
 from adc.api import AudioClassificationData
 
 
-class AdamsAudioClassificationWriter(SplittableStreamWriter, AnnotationsOnlyWriter, InputBasedPlaceholderSupporter):
+class AdamsAudioClassificationWriter(SplittableStreamWriter, AnnotationsOnlyWriter, InputBasedVariableSupporter):
 
     def __init__(self, output_dir: str = None, class_field: str = None, annotations_only: bool = None,
                  split_names: List[str] = None, split_ratios: List[int] = None, split_group: str = None,
@@ -65,7 +65,7 @@ class AdamsAudioClassificationWriter(SplittableStreamWriter, AnnotationsOnlyWrit
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output", type=str, help="The directory to store the audio/.report files in. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=True)
+        parser.add_argument("-o", "--output", type=str, help="The directory to store the audio/.report files in. Any defined splits get added beneath there. " + variable_list(obj=self), required=True)
         parser.add_argument("-c", "--class_field", metavar="FIELD", type=str, default=None, help="The report field containing the audio classification label", required=True)
         add_annotations_only_writer_param(parser)
         return parser
@@ -106,7 +106,7 @@ class AdamsAudioClassificationWriter(SplittableStreamWriter, AnnotationsOnlyWrit
         :param data: the data to write (single record or iterable of records)
         """
         for item in make_list(data):
-            sub_dir = self.session.expand_placeholders(self.output_dir)
+            sub_dir = self.session.expand_variables(self.output_dir)
             if self.splitter is not None:
                 split = self.splitter.next(item=item.audio_name)
                 sub_dir = os.path.join(sub_dir, split)