+++ src/adc/writer/speech/_hf_audiofolder.py
@@ -5,13 +5,13 @@ from typing import List, Iterable
 
 from wai.logging import LOGGING_WARNING
 
-from seppl.placeholders import placeholder_list, InputBasedPlaceholderSupporter
+from seppl.variables import InputBasedVariableSupporter, variable_list
 from kasperl.api import SplittableBatchWriter, AnnotationsOnlyWriter, add_annotations_only_writer_param
 from adc.api import SpeechData
 from adc.reader.speech import HF_AUDIOFOLDER_EXPECTED_HEADER
 
 
-class HuggingFaceAudioFolderSpeechWriter(SplittableBatchWriter, AnnotationsOnlyWriter, InputBasedPlaceholderSupporter):
+class HuggingFaceAudioFolderSpeechWriter(SplittableBatchWriter, AnnotationsOnlyWriter, InputBasedVariableSupporter):
 
     def __init__(self, output_dir: str = None, rel_path: str = None, annotations_only: bool = None,
                  split_names: List[str] = None, split_ratios: List[int] = None, split_group: str = None,
@@ -68,7 +68,7 @@ class HuggingFaceAudioFolderSpeechWriter(SplittableBatchWriter, AnnotationsOnlyW
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output", type=str, help="The directory to store the data. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=True)
+        parser.add_argument("-o", "--output", type=str, help="The directory to store the data. Any defined splits get added beneath there. " + variable_list(obj=self), required=True)
         parser.add_argument("--rel_path", type=str, help="The relative path to the audio files.", required=False, default=".")
         add_annotations_only_writer_param(parser)
         return parser
@@ -114,7 +114,7 @@ class HuggingFaceAudioFolderSpeechWriter(SplittableBatchWriter, AnnotationsOnlyW
         :type data: Iterable
         """
         for item in data:
-            sub_dir = self.session.expand_placeholders(self.output_dir)
+            sub_dir = self.session.expand_variables(self.output_dir)
             if self.splitter is not None:
                 split = self.splitter.next(item=item.audio_name)
                 sub_dir = os.path.join(sub_dir, split)