+++ src/adc/writer/speech/_piper.py
@@ -3,13 +3,13 @@ import os
 from typing import List, Iterable
 
 from kasperl.api import SplittableBatchWriter, AnnotationsOnlyWriter, add_annotations_only_writer_param
-from seppl.placeholders import placeholder_list, InputBasedPlaceholderSupporter
+from seppl.variables import InputBasedVariableSupporter, variable_list
 from wai.logging import LOGGING_WARNING
 
 from adc.api import SpeechData, FORMAT_WAV, FORMAT_EXTENSIONS
 
 
-class PiperSpeechWriter(SplittableBatchWriter, AnnotationsOnlyWriter, InputBasedPlaceholderSupporter):
+class PiperSpeechWriter(SplittableBatchWriter, AnnotationsOnlyWriter, InputBasedVariableSupporter):
 
     def __init__(self, output_dir: str = None, metadata: str = None, rel_path: str = None, speaker_key: str = None,
                  annotations_only: bool = None,
@@ -73,7 +73,7 @@ class PiperSpeechWriter(SplittableBatchWriter, AnnotationsOnlyWriter, InputBased
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output", type=str, help="The directory to store the audio/metadata files in. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=True)
+        parser.add_argument("-o", "--output", type=str, help="The directory to store the audio/metadata files in. Any defined splits get added beneath there. " + variable_list(obj=self), required=True)
         parser.add_argument("--metadata", type=str, help="The name of the meta-data file.", required=False, default="metadata.csv")
         parser.add_argument("--speaker_key", type=str, help="The key in the meta-data with the speaker name/ID; assumes single-speaker data if not supplied.", required=False, default=None)
         parser.add_argument("--rel_path", type=str, help="The relative path to the audio files.", required=False, default="wav")
@@ -125,7 +125,7 @@ class PiperSpeechWriter(SplittableBatchWriter, AnnotationsOnlyWriter, InputBased
         :type data: Iterable
         """
         for item in data:
-            sub_dir = self.session.expand_placeholders(self.output_dir)
+            sub_dir = self.session.expand_variables(self.output_dir)
             if self.splitter is not None:
                 split = self.splitter.next(item=item.audio_name)
                 sub_dir = os.path.join(sub_dir, split)