+++ CHANGES.rst
@@ -16,6 +16,7 @@ Changelog
 - the `convert-to-mono` filter now has a `-f/--force` flag to enforce the conversion
 - added the `change-volume` filter
 - added support for data-formatters with `df-audio-data` specifically for audio data
+- using `seppl.variables` now
 
 
 0.1.0 (2025-10-31)